[linux] fix integer overflow in free/available memory info on 32bit
The sysinfo members are unsigned long, cast them to uint64_t before
calculating the size in bytes to prevent 32bit integer overflow
on 32bit systems with 4GB or more RAM.

+++ xbmc/platform/linux/XMemUtils.cpp
@@ -167,10 +167,10 @@ void GlobalMemoryStatusEx(LPMEMORYSTATUSEX lpBuffer)
     fflush(procMeminfoFP);
   }
   lpBuffer->dwLength        = sizeof(MEMORYSTATUSEX);
-  lpBuffer->ullAvailPageFile = (info.freeswap * info.mem_unit);
-  lpBuffer->ullAvailPhys     = ((info.freeram + info.bufferram) * info.mem_unit);
-  lpBuffer->ullAvailVirtual  = ((info.freeram + info.bufferram) * info.mem_unit);
-  lpBuffer->ullTotalPhys     = (info.totalram * info.mem_unit);
-  lpBuffer->ullTotalVirtual  = (info.totalram * info.mem_unit);
+  lpBuffer->ullAvailPageFile = (static_cast<uint64_t>(info.freeswap) * info.mem_unit);
+  lpBuffer->ullAvailPhys     = (static_cast<uint64_t>(info.freeram + info.bufferram) * info.mem_unit);
+  lpBuffer->ullAvailVirtual  = (static_cast<uint64_t>(info.freeram + info.bufferram) * info.mem_unit);
+  lpBuffer->ullTotalPhys     = (static_cast<uint64_t>(info.totalram) * info.mem_unit);
+  lpBuffer->ullTotalVirtual  = (static_cast<uint64_t>(info.totalram) * info.mem_unit);
 #endif
 }
